frontend/requirements.txt
pandas
matplotlib
fdb
python-dotenv
streamlit
openpyxl
